+++ braintree/exceptions/http/timeout_error.py
@@ -2,3 +2,11 @@ from braintree.exceptions.unexpected_error import UnexpectedError
 
 class TimeoutError(UnexpectedError):
     pass
+
+
+class ConnectTimeoutError(TimeoutError):
+    pass
+
+
+class ReadTimeoutError(TimeoutError):
+    pass